deploy capo

+++ controllers/users.js
@@ -1,5 +1,6 @@
 const User = require('../model/users');
 const jwt = require('jsonwebtoken');
+const axios = require('axios');
 require('dotenv').config();
 const claveToken = process.env.CLAVE;
 const bcrypt = require('bcrypt');
@@ -11,7 +12,6 @@ const getUser = async (req, res) => {
     } catch (error) {
         res.status(500).json({ error: error.message });
     }
-
 }
 
 const getUserEspecifico = async (req, res) => {
@@ -21,8 +21,10 @@ const getUserEspecifico = async (req, res) => {
 }
 
 const crearUser = async (req, res) => {
-    const { username, name, surname, email, password, rol } = req.body;
-
+    const { username, name, surname, email, password, role } = req.body;
+    const credits = 0;
+    const date = new Date();
+    const expire = new Date(date.getFullYear(), date.getMonth() + 1, date.getDate());
     const saltRound = 15; 
     const passwordEncripted = bcrypt.hashSync(password, saltRound);
     const nuevoUser = new User({
@@ -31,7 +33,9 @@ const crearUser = async (req, res) => {
         surname,
         email,
         password: passwordEncripted,
-        rol
+        credits,
+        expire: expire.toLocaleDateString('es-ES'),
+        role
     })
     await nuevoUser.save()
     res.status(200).send(`Se creo el usuario con éxito.`)
@@ -39,29 +43,58 @@ const crearUser = async (req, res) => {
 
 const deleteUser = async (req, res) => {
     const { id } = req.body
-    await User.findByIdAndDelete(id);
-    res.status(200).send(`Se elimino el usuario con éxito.`)
+    if (id) {
+        await User.findByIdAndDelete(id);
+        res.status(200).send(`Se elimino el usuario con éxito.`)
+    } else{
+        res.status(500).send(`No id.`)
+    }
+
 }
 
+const takeCredit = async (req, res) => {
+    const { id, credits } = req.body
+    await User.findByIdAndUpdate(id, {
+        credits,
+    })
+    res.status(200).send(`Se resto el crédito con éxito.`)
+};
+
+const removeCredits = async (id) => {
+    const credits = 0;
+    await User.findByIdAndUpdate(id, {
+        credits,
+    })
+};
+
+const addCredits = async (req, res) => {
+    const { id, credits } = req.body
+    const date = new Date();
+    const expire = new Date(date.getFullYear(), date.getMonth() + 1, date.getDate());
+    await User.findByIdAndUpdate(id, {
+        credits,
+        expire: expire.toLocaleDateString('es-ES')
+    })
+    res.status(200).send(`Se actualizaron los créditos con éxito.`)
+};
+
 const patchUser = async (req, res) => {
-    const { id, username, name, surname, email, password, rol } = req.body
-
-
+    const { id, username, name, surname, email, password, role } = req.body
     await User.findByIdAndUpdate(id, {
         username,
         name,
         surname,
         email,
         password,
-        rol
+        role
     })
     res.status(200).send(`Se actualizo el usuario con éxito.`)
 };
 
 const rolUser = async (req, res) => {
-    const { id, rol  } = req.body
+    const { id, role  } = req.body
     await User.findByIdAndUpdate(id, {
-        rol
+        role
     })
     res.status(200).send(`Se actualizo el usuario con éxito.`)
 };
@@ -116,4 +149,19 @@ const restablecerContraseña = async (req, res) => {
     res.status(200).send(`Se actualizo la contraseña con éxito.`)
 };
 
-module.exports = { crearUser, getUser, deleteUser, patchUser, getUserEspecifico, rolUser, loginUser, emailUser, restablecerContraseña }
+const checkUsers = async () => {
+    const users = await User.find({})
+    const now = new Date();
+    for (const user of users) {
+      const expireDate = new Date(user.expire.split('/').reverse().join('-'));
+      if (now > expireDate) {
+        await removeCredits(user.id);
+      }
+    }
+  };
+
+  setInterval(() => {
+    checkUsers();
+  }, 5 * 60 * 1000);
+
+module.exports = { crearUser, getUser, deleteUser, patchUser, getUserEspecifico, rolUser, loginUser, emailUser, restablecerContraseña, takeCredit, addCredits }
